refactor(operate): introduce OperatorSinkPlugin seam; extract TV bits to sinks/broadcast.ts
First step of the operator-driver factoring: introduce a domain-plugin
seam in the operator daemon and move all SportsClaw-TV-specific code
behind it. Sportsclaw core stays generic; the broadcast sink (and the
recall + generate_image hookups + tail-server poster + tv-content-
archive writer that came with it) consolidates into one place that's
ready to ship out to the machina-sports-tv repo in the follow-up.

## Why

src/operate.ts had accumulated TV-specific opinions across #46 / #48 /
#49 / #50 / #52:

  - <<<DATA>>>...<<<END>>> packet protocol (broadcast invention)
  - mapTickEventToTelemetry kind vocabulary (broadcast/tick/reasoning/
    error/gate — overlay vocab)
  - tail-server URL convention (TV-specific endpoint)
  - tv-content-archive doc name + category enum (TV-specific schema)
  - tv-content-library doc + scope_type enum (World Cup framing)
  - broadcaster-flavored generate_image description + sink

A betting / scouting / fan-engagement operator built on the daemon
today would inherit all of those and either fork the file or write
betting data to a misleadingly-named tv-content-archive doc. The
session-1 prompts refactor (#44) already domain-neutralized the prompt
fragments; this is the same kind of work for the driver.

## What lands in this PR

  src/operator-sink.ts              NEW  interface + noopSink + resolver
  src/sinks/broadcast.ts            NEW  all TV-specific code, behind
                                          the plugin interface
  src/operate.ts                    -358 / +29  collapses to a thin
                                          driver that routes events
                                          through the resolved sink
  src/operator-config.ts            +sink?: string field + validation
  test/operate.test.mjs             trimmed to generic CLI tests only
                                    (parseFlags, exitCodeFor, resolve-
                                    Persona, resolveExtraFragments)
  test/sinks-broadcast.test.mjs     NEW  all the TV-specific tests
                                    (poster envelope, packet strip,
                                    recall+generate_image registration)
  test/operator-sink.test.mjs       NEW  resolver + noopSink pins
  test/operator-config.test.mjs     +5  new tests for the sink field

## The seam

```ts
export interface OperatorSinkPlugin {
  name: string;
  registerTools?(args: {toolSet, toolNames, cfg, mcpManager}): void;
  wrapImageGenerator?(args: {cfg, mcpManager}): CreateGenerateImageToolOpts;
  onTickEvent?(evt: TickEvent, ctx: SinkContext): Promise<void> | void;
  onToolCall?(evt: ToolCallEvent, ctx: SinkContext): Promise<void> | void;
}

export interface SinkContext {
  jobId: string;
  modelId?: string;
  intervalMs?: number;
  mcpManager?: McpManager;
  cfg: OperatorJobConfig;  // sinks read domain-specific fields here
}
```

## Resolver

`resolveSink(cfg)` returns a sink based on:

  1. cfg.sink === "noop"      → noopSink
  2. cfg.sink === "broadcast" → bundled broadcast sink (this PR keeps it
                                in sportsclaw for backward compat; it
                                moves to machina-sports-tv next PR)
  3. cfg.sink === <other>     → throws "not a built-in" (external sink
                                resolution lands when TV ships its pkg)
  4. cfg.tailServer is set    → backward compat: broadcast sink
  5. otherwise                → noopSink

## Backward compatibility

Existing operator job configs work unchanged:
  - Configs with `tailServer` resolve to the broadcast sink (case 4).
  - Configs with neither field get noopSink, which prints TickEvents to
    stdout — matching the pre-refactor "no tailServer" behavior.

Verified with the live tv-operator config:
  - --dry-run resolves "sink: broadcast", lists generate_image +
    recall_recent_content + recall_library in the 298-tool set
  - --once tick succeeds: 6 tool calls, well-formed broadcast + packet
    (3 markets / 3 fixtures / 3 news), telemetry POSTs to the tail-
    server, archive writes to the pod — all behaviors preserved

## Tests

432 pass (was 420, +12 from new sink tests). Build clean.
