+++ package.json
@@ -43,7 +43,9 @@
     "test:ci-integration": "npm run build && node test/ci-integration.mjs",
     "test:pod-memory-append-race": "npm run build && node --test test/pod-memory-append-race.test.mjs",
     "test:halt-guard": "npm run build && node --test test/halt-guard.test.mjs",
-    "test:tool-call-part-guard": "npm run build && node --test test/tool-call-part-guard.test.mjs"
+    "test:tool-call-part-guard": "npm run build && node --test test/tool-call-part-guard.test.mjs",
+    "test:operator-sink": "npm run build && node --test test/operator-sink.test.mjs",
+    "test:sinks-broadcast": "npm run build && node --test test/sinks-broadcast.test.mjs"
   },
   "keywords": [
     "sports",
